reorgnization, more fixes
git-svn-id: svn://svn.macaulay2.com/Macaulay2/trunk@7709 a8012f80-5f13-0410-8dc0-eb77bd940c0d

+++ M2/Macaulay2/d/M2mem.h
@@ -1,12 +1,14 @@
+#include <stddef.h>
+
 #if defined(__cplusplus)
 extern "C" {
 #endif
   extern void outofmem(void);
-  extern char *getmem(unsigned int);
-  extern char *getmem_clear(unsigned int);
-  extern char *getmem_atomic(unsigned int);
-  extern char *getmem_malloc(unsigned int);
-  extern char *getmem_atomic_clear(unsigned int);
+  extern char *getmem(size_t);
+  extern char *getmem_clear(size_t);
+  extern char *getmem_atomic(size_t);
+  extern char *getmem_malloc(size_t);
+  extern char *getmem_atomic_clear(size_t);
 #if defined(__cplusplus)
 }
 #endif
